incorporate meterValues
- Added element "meterValue" to InformProvider.req
- Added element "localTime" to InformProvider.req

This is to enable kWh-based billing. In this case, for "start" and "end" messages sent out by the CPO, they are to contain localTime and meterValue information to be displayed to the customer at the start and at the end of their charging process to enable them to check against the meter itself and against their final bill.