Example Endpoints for OCHPdirect (as suggested by regioIT)
regioIT needs defined ports for all bindings to make the WSDL work in their toolset. As all ports are changed after implementation anyway and overwritten with productive versions instead of the WSDL-defined ones, we could add these non-functional example ports to the WSDL to enable other implementing partners using the same or a similarly behaving toolset to properly import the WSDL including the Provider and Operator Endpoint operations for OCHPdirect.